Password field added to Artifactory User
Terraform can store the "known" state of a password and update it only
when the input is updated. The password is hashed so it is not stored in
plain text but can be compared for changes.

While import will fail to get the password initially, this can be
ignored using other Terraform primitives.