User password converted to be optional
In doing this the password gets automatically set when a new user is
created and no password is explicitly set. When a password is set, it is
then set to that. Removing the password field removes "management" of
the password while leaving it's value in tact.

This implementation still does not deal with changes upstream from the
server itself, but also will not overwrite changes made on the
Artifactory server.